Extract shared FakeChatClient
