+++ ConsoleChat.Tests/ChatControllerTests.cs
@@ -1,34 +1,15 @@
-using System.Runtime.CompilerServices;
 using Microsoft.Extensions.AI;
 using NSubstitute;
 using SemanticKernelChat;
 using SemanticKernelChat.Console;
 using SemanticKernelChat.Infrastructure;
 
+using ConsoleChat.Tests.TestUtilities;
+
 namespace ConsoleChat.Tests;
 
 public class ChatControllerTests
 {
-    private sealed class FakeChatClient : IChatClient
-    {
-        public ChatResponse Response { get; set; } = new(new ChatMessage(ChatRole.Assistant, "reply"));
-        public List<ChatResponseUpdate> StreamingUpdates { get; } = new();
-
-        public Task<ChatResponse> GetResponseAsync(IEnumerable<ChatMessage> messages, ChatOptions? options = null, CancellationToken cancellationToken = default)
-            => Task.FromResult(Response);
-
-        public async IAsyncEnumerable<ChatResponseUpdate> GetStreamingResponseAsync(IEnumerable<ChatMessage> messages, ChatOptions? options = null, [EnumeratorCancellation] CancellationToken cancellationToken = default)
-        {
-            foreach (var update in StreamingUpdates)
-            {
-                yield return update;
-                await Task.Yield();
-            }
-        }
-
-        public object? GetService(Type serviceType, object? serviceKey) => null;
-        public void Dispose() { }
-    }
 
     [Fact]
     public async Task SendAndDisplayAsync_Adds_To_History_And_Writes_To_Console()
